feat: create source folder and move app
