+++ src/App.tsx
@@ -11,8 +11,6 @@
 import React from 'react';
 import {SafeAreaView, StatusBar} from 'react-native';
 
-declare const global: {HermesInternal: null | {}};
-
 const App = () => {
   return (
     <>
